Test OpenGl methods in FakeBatcher test
Part of the work to split Batcher in BatcherInterface and
PrimitiveAssembler.
(https://github.com/karupayun/orbit/commit/1d4d264fc5f1f90c033b62bc2d0b22f45750401f).

This class is testing different things (internal implementation, picking
and shared pickable) that now are going to be in separate files. As a
first step I'm testing the internal methods that are going to be used in
OpenGl implementation instead of the PrimitiveAssembler part.

In addition I'm renaming the internal methods to the original name since
they are going to be public functions used by the PrimitiveAssembler
(Henning's suggestion).

As a side effect, CaptureWindow complains correctly that they expect
that methods to be in OpenGlBatcher. This is because CaptureWindow
should use primitive assembler there. This is already in
https://github.com/karupayun/orbit/commit/1d4d264fc5f1f90c033b62bc2d0b22f45750401f,
but for now I just add a temporal TODO.

- As little temporal detail, I add GetBatcherId() and GetNumElements()
  temporary to Batcher.

2nd step in http://b/225173189