Create MockBatcher class
Designed in go/stadia-orbit-testing-ui-elements. In this PR we are
created MockBatcher class. This class is going to be used to test
UpdatePrimitive and Draw methods but it also will be useful to test
PrimitiveAssembler class, so I'm creating it before its renaming so the
new class can be properly tested.

http://b/223559986