Support anonymously mapped PE/COFF files in LinuxMap
We want to detect whether an anonymous mapping corresponds to the .text segment
of a PE loaded by Wine. The mapping is anonymous if the offset of the section in
the file (`PointerToRawData`) is not congruent to the offset when loaded into
memory (`VirtualAddress`) modulo the page size.

The idea is the following:
Given an anonymous executable mapping, consider the previous "group" of file
mappings that are all backed by the same file, and in particular consider the
first of these mappings. If that file is a PE, and there is no executable file
mapping in this "group" of file mappings, and the address range where the .text
section of that PE should be mapped (based on the file mapping's address and on
the .text section's `VirtualAddress` and `VirtualSize`) is contained in the
anonymous mapping's address range, then we can be fairly sure that the anonymous
mapping does indeed correspond to the .text of the PE.

We add this logic to `orbit_object_utils::ParseMaps` so that PEs with the
characteristic mentioned above are recognized as modules and loaded. The class
`FileMappedIntoMemory` helps contains most of the logic and allows us to perform
the detection while scanning `/proc/[pid]/maps` line by line.

More information on the problem in the documents linked in the bug.

Bug: http://b/226562022

Test: Unit tests. Load modules of unaligned `triangle.exe` and of Virtus: verify
the main binary shows up. Use user space instrumentation on `triangle.exe` and
on Virtus: this confirms addresses are correct.