Create MockBatcher class (#3525)
* Create MockBatcher class

Designed in go/stadia-orbit-testing-ui-elements. In this PR we are
created MockBatcher class. This class is going to be used to test
UpdatePrimitive and Draw methods but it also will be useful to test
PrimitiveAssembler class, so I'm creating it before its renaming so the
new class can be properly tested.

http://b/223559986

* Fixup: Add public ExpectDraw instead of GetNumElement methods

* Revert "Fixup: Add public ExpectDraw instead of GetNumElement methods"

This reverts commit b01d84b712356091d0ac12b16854ace98b811ab3.

* Fixup: Addressing comments

* Fixup: Refactoring methods

* Fixup: Addressing comments

* Fixup: clang-format

* fixup: Static