Add utility function to test if a file is potentially a PE/COFF file
This is needed to distinguish between PE/COFF and ELF files during
initialization of object files when they are first encountered during
unwinding. The function doesn't have to (and can't) correctly predict
if an object file is really a PE/COFF (unless we'd fully read, parse,
and validate everything, which is not feasible), but there are later
safeguards that prevent incorrect handling when the file is not actually
a PE/COFF file. An important property is however that an ELF file is not
incorrectly detected as a PE/COFF file.

Tested: Unit tests.
Bug: http://b/194768602