Expand functionality for detecting and handling function epilogs
This change mainly pulls in some of the surrounding logic for
detecting and handling epilogs of functions into the PeCoffEpilog class,
simplifying the general unwinder code (yet to be added).

Instead of just taking a vector of bytes representing machine code, the
DetectAndHandleEpilog method will now read the machine code from the
object file at the given location and then evaluate the machine code for
being an epilog.

Tested: Unit tests pass.
Bug: http://b/194768602