Split Batcher with PrimitiveAssembler
In this PR we are finally splitting Batcher with PrimitiveAssembler
(step 2 in http://b/225173189).

This is a big PR that consist in several little things. I wanted to
make it in few smaller steps (https://github.com/google/orbit/pull/3517)
but turned to be that it was difficult to understand the big picture
and we agreed with Henning in making everything in one PR.

TimeGraph owns now a PrimitiveAssembler and a Batcher. The first is
passed to CaptureViewElements to draw and the second is used by
CaptureWindow to draw. CaptureWindow also owns another copy (at it was
before). This is a bit confused (actually one batcher is used to
UpdatePrimitives(timers) and the other for the rest of the elements),
but it was like that before and I'm not aiming to solve this.

OpenGlBatcher is adapted to Batcher/BatcherInterface.

PrimitiveAssembler is simplified and now depends on having a Batcher
to perform additions. There is no drawing logic in this class anymore,
ResetElements neither Translations.

Fixed an small issue in AddCircle (we were making double-translations)

BatcherTest
- Now is faking OpenGlBatcher and it will renamed in a following PR to
OpenGlBatcherTest (http://b/228053356)
- Erased 2 tests (PickingElementsDrawing is fully replaced in
PrimitiveAssemblerTest and PickingPickables is part of http://b/228063067).