Show capture length in the load capture widget (#3493)
* Add capture length to CaptureFileInfo

With this change, we added `capture_length` into `CaptureFileInfo`.

Bug: http://b/198562119

* Add a capture length column to ItemModel

With this change, we add a "Capture length" column to the load capture
widget. If the capture length is unknown (e.g., previous files that we
haven't touched yet), we show "--" to indicate that the capture length
information is still missing. Also we will add a tooltip for this later.

Bug: http://b/198562119

* Add capture length info when loading a capture successfully

With this change, we add the capture length info when we load a capture
file successfully.

Bug: http://b/198562119

* Add capture length info when moving / renaming a capture file

With this change, we added the capture length info when moving or
renaming the capture file.

Bug: http://b/198562119

* Add capture length info after taking the capture

With this change, we changed to call `AddOrTouchCaptureFile` on capture
finished rather than on capture started. This makes us able to get the
capture length info after taking a capture.

Bug: http://b/198562119

* Update tests for CaptureFileInfo Manager & ItemModel

With this change, we update the unit tests for CaptureFileInfo Manager
and ItemModel.

Bug: http://b/198562119

* Add tooltips for missing capture length info

With this change, we added the tooltips for the missing capture length info to
indicate that the information will be available once the capture file is loaded.

Bug: http://b/198562119

* [fixup] Address review comments

* [fixup] Add capture length to CaptureFileInfo constructor

* [fixup] Replace kMissingCaptureLengthValue by using optional

* Save file size and last modified time

With this change, we:
* Add and save `CaptureFileInfo::file_size_` and `last_modified_`
* Enable `CaptureFileInfo` to check whether the capture file information
  are out of sync. This enable us to check whether a capture file is
  modified outside Orbit
* Let the manager reset capture length if the `CaptureFileInfo` is out
  of sync

Test: Unit test; Modify a capture file name from the file explorer,
check that the capture length value in the load capture wideget becomes
"--".
Bug: http://b/227547320, http://b/227554942

* [fixup] Address review comments