Expose "stack dump size" in capture options (devmode)
For frame pointer unwinding, we require to copy a small
portion of the stack to fix leaf functions. It is essential
for performance that this value is small. However, sometimes
the default value is too small, and the unwinding errors are
too many.

In those cases, we want to be able to adjust the size of the stack
to copy.

This change exposes the option via the capture option dialogue
(in devmode). For now, we only allow this for frame pointer
unwinding. The maximal value allowed to use is 65000, while the
default value is 512.

Note, in the Dwarf case, we will remain using the current default
of 65000.
Further note, we are now not anymore able to set the value for
Dwarf. However, this is fine, as the specified value is always
an upper bound. If the actual stack is smaller, only the actual
size will be copied. For Dwarf, we will always require the complete
stack.

This also removes the flag to specify the max stack size to copy.

http://screenshot/35CX69b2MeUbmqA
http://screenshot/5PdoM6JaGa3tk4W

Bug: http://b/228187977
Test: Manual test capture options ui with different sampling modes
      and values for the text field. Take captures with different
      unwinding methods and max stack dump sizes; verify unwinding
      errors.