Add PeCoff::GetTextRange, PeCoff::GetTextOffsetInFile to libunwindstack
These will be used to adjust `unwindstack::Maps` in order to support unwinding
through frames of functions in Portable Executables whose executable section is
mapped anonymously in `/proc/[pid]/maps` by Wine because the file alignment
doesn't obey the requirements of `mmap`.

`GetTextRange` was already available with the same semantics for `Elf`, and is
now made a `virtual` method of `Object`. `GetTextOffsetInFile` is only added to
`PeCoff`.

Bug: http://b/226562022

Test: Unit tests.