Make zooming-out in the histogram always react
Previously it was possible to stack up multiple equal selections.
In that case a single click (that is supposed to zoom-out)
was switching from the current histogram to the previous one
that was exactly the same. This was confusing to the user.

A simple check is added to avoid such stacking-up.

Tests: Manual
Bug: http://b/228167388