Improve PrimitiveAssemblerTest
In this PR I'm improving the test in PrimitiveAssemblerTest to test all
the possible additions for the PrimitiveAssembler.

For all the basic and combination methods we are testing the number of
calls we do to the Batcher and also if everything is inside the
rectangle that it should be.

Test are pretty basic though, I'm always using the same coordinates but
it should be enough.