Move FindModuleLocally off the main thread
This changes OrbitApp::FindModuleLocally to return type Future and
changes the functionality to use thread_pool when searching for local
files. This is done because in rare cases this find operation can take
longer than expected and then blocks the main thread. This is not
solving the root cause that but rather mitigates the main thread issue.

Bug: http://b/228599959
Test: Manual test that symbol loading of local files still works