Make 2D shapes store Vec2, not Vec3
Previously, Tetragon, Line and Traingle stored their vertices
in Vec3, thus replicating the z-component corresponding to
the level. Equality of the z-component for all the vertices
was not enforced.

To acces the z-component we used to write `shape.vertices[0][2]`.

This makes the said shapes to store only x- and y-components,
the z-component is now passed as a separate variable.

As GL expects 3d-vectors, we have to eventually transform
the Vec2-vertices to Vec3 vertices, thus introducing
Tetragon3D, Line3D and Traignle3D structs, that are
hidden from the caller.

This also fixes two bugs that became apparent during the
rafactoring. Bugs like these are much less likely to happen
with new API.

1. `OpenGlBatcher::AddLine` was translating the Line,
including the z-components. Yet, the translated line
(storing the translated z-component) was placed onto
the layer corresponding to non-translated z-component.
This fixes that along with the buggy test
`TEST(OpenGlBatcher, TranslationsAreAutomaticallyAdded)`.

2. `Track::DrawTriangleFan` was placing the triangles on the
layer corresponding to 2*z, not z.

Test: Compile, Unit, Manual
Bug: http://b/227748244