Implement ObjectFile::GetExecutableSegmentSize (#3550)
This is the counterpart of `GetExecutableSegmentOffset`. Both refer to the
executable segment when it's loaded into memory.

It will be necessary for the detection of Windows binaries whose executable
segment is mapped anonymously in `/proc/[pid]/maps` by Wine because the
`FileAlignment` is not multiple of the page size.

Bug: http://b/226562022

Test: Unit tests. Tried as part of a larger prototype for the problem mentioned
above.