Create DrawingInterface class
In this PR we are creating DrawingInterface class. We will move
GetLayers, DrawLayer and Draw methods from Batcher and TextRenderer
there (implemented by OpenGl methods).

Also, we will implement it in CaptureWindow with a new class
DrawingCombiner which will take care of combining Layers (logic which is
currently in CaptureWindow::Draw()