Create MockTimelineInfo
In this PR we are creating MockTimelineInfo. It's a simpler
implementation than the one in TimeGraph and will be used for making
unit tests for CaptureViewElements. For example it's needed in every
Track since Track ask for a TimelineInfo when they are constructed.

The MockTimelineInfo simply assumes that capture start at timestamp = 0
and provides methods to modify Min/Max visible timestamps (which also
updates the MaxCaptureTimestamp). Also provides conversions between
Tick/Us/World using internal width.

http://b/228610504