Move TextRenderer to orbit_gl namespace
In this PR we are moving TextRenderer to orbit_gl namespace. We are
planning to Mock that class and since the mock class, the interface and
TextRenderer itself should be in that namespace we are previously
solving that tech-debt.

In addition we are erasing 3 forward declarations of TextRenderer that
are not needed anymore.

http://b/228609901

Test: The code compiles.