Detect if frame pointer register is set without unwinding
We use Dwarf unwinding to patch missing callers of leaf functions.
In case we are not in a leaf function, we don't need to do
anything.
When the collected stack size is too to perform unwinding, we
still return an unwinding error, even for non-leaf functions.
However, in this case, the Dwarf information will tell us
that the frame pointer is actually set correctly.

So in this change, we make use of this fact and read the
debug frame information by ourselves.

Bug: http://b/228444252
Test: Run tests & check unwinding errors on various samples