Move PE/COFF unwind info evaluation into its own class for testing
This change renames the general PE/COFF unwinder class to a specific
class only intended for evaluating unwind infos. This class will later
be used by the general unwinder class as part of the 'Step'
implementation.

This change also adds a mapping of relative virtual addresses (RVA) to
file offset directly inside the PeCoffUnwindInfos class to make usage
of that class easier for external code. This also minimizes usage
of the function to map RVA to file offsets, which can be expensive,
by caching unwind info entries based on the RVA and not the file offset.

Tested: Unit tests.
Bug: http://b/194768602