Replace fake track header with real element (#3491)
We're currently rendering the track as a single element, but expose it as a
composition of track, tab and timer pane in E2E tests. As a first step to fix
this, the tab is now converted to an actual CaptureViewElement.

The new TrackHeader expects to be parented under a Track, and construction
will fail if the parent is not of type Track. This allows us to directly interact with
the parent through the more specialized Track API.

While this to some extent breaks the generic parent / child relationship, it is
IMO a common restriction that certain UI elements have to live as children of
specific parents, and I think it's an acceptable solution. Let me know if there's
other ideas.

Bug: b/185854980
Tests: E2E test "orbit_tracks" works, and manual testing.