Execute layout changes in a loop
Layout changes of one element may require other elements to be updated
as well, so layouting needs to be done until all elements report that
they do not need to be updated further. As layout requests bubble up,
it's enough to check this for the root element (time graph) of the tree.

During loading or capturing, only a single layouting loop is
executed as we're streaming in data from a seperate thread.

Bugs:
b/227989316
b/227162509

Test: Manual testing