Defer deletion of frame tracks.
This could be a potential source of issues: Frame tracks are being
deleted, but the pointer is still referenced in visible_tracks_. Update
of visible_tracks_ will eventually happen during UpdateLayout(), but
until then, we keep the stale pointer in this array.

Even though deletion of tracks happens from the main thread, it could
still happen in between UpdateLayout() and Draw() when they are handled
during different executions of the main thread.