Make Hooked column under Live tab show entry types
Now we include manual instumentation under Live tab,
so we have to also report the types of the entries.
This is done via reporting the abbreviations e. g.
MS means manual, syncronous.

As the tab now reports much more than hook-edness
(it already was showing whether frametrack is enabled,
but now we add types), its name is also changed.

Further, as the column is not called "Hooked" anymore,
the tick sing is not self-explanatory, so this also
replaces the tick with `H`.

Test: Manual, Unit, E2E
Bug: http://b/229218555