Unify updates of TimeGraph time info (#3584)
This fixes the issue that double-clicking on an event does not correctly
update the capture window. The reason is that we set the capture
min / max times (both for the full capture and the visible area) in
multiple places. Not all of them trigger a RequestUpdate, and in this
particular case, the values for min / max time may have been updated
before redrawing, but the visible_timeframe_ may not.

This PR unifies all modifications to min and max timestamps and makes
sure we call RequestUpdate on all of them.

Bug: b/227162509
Test: Manual testing