Make 2D shapes store Vec2, not Vec3
Previously, Tetragon, Line and Triangle stored their vertices
in Vec3, thus replicating the z-component corresponding to
the level. Equality of the z-component for all the vertices
was not enforced.

To access the z-component we used to write shape.vertices[0][2].

This makes the said shapes to store only x- and y-components,
the z-component is now passed as a separate variable.

This also makes GL to accept 2-dimensional vertices and
disables depth-testing.

This also fixes two bugs that became apparent during the
refactoring. Bugs like these are much less likely to happen
with new API.

OpenGlBatcher::AddLine was translating the Line,
including the z-components. Yet, the translated line
(storing the translated z-component) was placed onto
the layer corresponding to non-translated z-component.
This fixes that along with the buggy test
TEST(OpenGlBatcher, TranslationsAreAutomaticallyAdded).

Track::DrawTriangleFan was placing the triangles on the
layer corresponding to 2*z, not z.

Also, this renames Tetragon into Quad. 

Test: Compile, Unit, Manual
Bug: http://b/227748244
Bug: http://b/172557467