Create visual tests in the Timeline
In this PR we are creating visual tests in the TimelineUi component.
After creating several mock classes, we were able to test comparisons
between number of major ticks, number of minor ticks and labels for
different ranges. We were also able to test number of characters in the
labels, and positions of objects.

As a result, one bug was found and fixed (the margin below timeline was
outside of the kZTimeline ranges and on top of the incomplete data
intervals: http://screenshot/56bT7Cr7b5aE5GM.

Also, we found something we previously now (small screens works a bit
weird). This is a small issue that need a bit of a refactor to be fixed
and not a big priority but it was good that it was catched in
these unit tests. The issue is that screens below 700 pixels width
sometimes has intersecting labels (actually it's more like 500 but the
MockTextRenderer is estimating every character as a 'w').

Finally, we don't expect in really small screens to force to have at
least 2 labels.

Bug: http://b/218302149.