Patch leafs even when stack dump is too small
Previously, we gave up patching leaf functions, if the collected
raw stack was too small to include both the leaf's and the caller's
frame completely.

However, if the stack dump is only missing parts of the caller's
frame, unwinding may still be successfull to find out the leaf's
caller. In this case, we can still patch the callchain.

Now, we only return an unwinding error, if we actually not receive
two frames and the stack size is too small.

Test: Run tests & manually verify on samples.
Bug: http://b/228445173