Skip /dev/ files in detection of anonymously mapped PE/COFF files (#3596)
Calling `CreateObjectFile` hangs on `llvm::object::ObjectFile::createObjectFile`
for character or block devices. So don't call it on files in `/dev/`.
An alternative might be to use `std::filesystem::is_regular_file`, but I decided
to do the same as `CreateModule` at the top of `LinuxMap.cpp`.

Bug: http://b/229706955

Test: User space instrumentation on a few runs of Trata, that was previosly
showing the issue.