Process launching on Windows
We can now launch a process Windows and optionally pause it at entry
point and resume it on capture start. The feature is only enabled when
launching with "--local" and "--devmode".

https://screenshot.googleplex.com/95pLBxXXL7yYCAA.png

This change contains 3 main parts:

1. Creation LaunchedProcess class which holds client-side information
about a processe launched on the service-side. It also manages state
transitions from "spinning" to "suspended" to "running". The launched
processes are themselves mananged by the ProcessManager.

2. Adding MockProcessClient to be able to test the LaunchedProcess class.

3. Adding the grpc services to link client side requests to service side
actions which use the existing "ProcessLauncher" to perform the actual
process launching.

If this is too large a PR, I could split it, but I feel it might be
easier to get the full picture this way. It looks big but is actually
less than 2/3 what it looks like:

The 3 files in WindowsUtils are trivial renames, we have 2 CMakeLists
files and 2 .proto files, leaving 12 files for the actual core change.