Fix hooking a function when a capture is not connected
Currently, when no capture is connected and the user
enables a frametrack for a dynamically instrumented function,
the Type column of the Live tab shows the `F` character
(which is correct) and `H` character, which makes no sense,
as with no capture connected, the function cannot be hooked
on the next capture as the next capture is not possible.

Tests: Unit, Manual.
Bug: http://b/229709235