Fix permission issue in E2e test.
The capture files used in "capture loading" e2e test are now copied to
a temporary loaction before we open them. This is done since we need
write access to these files (see http://b/229836007).

This problem surfaced when Orbit was started with different permissions
in the e2e tests (see http://b/229417627).

Test: Reproduced the issue locally and ran the fixed script in the same
      environment.