Expose "potentially broken" frames
On unwinding errors in the service, we cut off the potetially
broken call frames and only report the PC, which we are pretty
sure to be correct.

However, the "broken" frames may contain important information.
e.g. in the case of frame pointer unwinding, they might only miss
one frame.

With this change, we do not cut "broken" frames, but report them
and show them in orange.
We also highlight the "Unwind error" string in the label.

Bug: http://b/230082209
Test: Take a capture