Fix leaf function patching when requiring red zone (#3615)
Libunwindstack will remove "speculative" frames, when
two unwinding errors in a row occurs.
There are special situations (ret instruction), where
the unwind information requires reading from the red-zone
(rbp in that case), but unwinding actually works perfectly
fine (as rbp is actually not modified). For full unwinding,
libunwindstack will recover and actually completely report
the callstack. However, for leaf function patching, we
only passed in rbp-rsp, so unwinding the second frame will
fail. Libunwindstack will assume the "speculative" frame
was incorrect and removes this.
For the leaf function patching this means, we receive only
one frame from libunwindstack, thus, no need to patch.

This change fixes this issue, by taking a completely new
approach to leaf function patching:
We only unwind one frame from libunwindstack (i.e. the frame
with the instruction pointer), but also read the updated
registers after that unwinding step.
In particular, if rbp was not changed during unwinding,
we know, this was a leaf function, and we can patch in
the updated "pc" entry from the registers, which contains
the "return address".
If, rbp was changed, this could either be an unwinding error,
or a function with frame pointers. We retain our heuristics
to detect unwinding errors and actually send unwinding
errors when non-executable code is on the callstack.

To be still able to handle cases, where the collected
callstack is too small, to actually unwind, but we know
that the sampled function has a frame pointer, we keep
the HasFramePointerSet functionality.

Unfortunately, libunwindstack does not report the canonical
frame address. Otherwise, we would have been able to actually
tell if the function had frame pointers set (there could
be cases were CFA does not get computed by $rbp + 16, but
the evaluated value is still $rbp+16).

A potential follow-up would be to improve HasFramePointerSet
such that it computes CFA instead of matching the computation
rule.

Bug: http://b/228428930
Test: Run tests & manually inspect callstacks in samples