Refactor SftpCopyToLocalOperation
This refactors SftpCopyToLocalOperation to use slightly different named
and ordered states. The states kNoOperation and kRemoteFileOpened are
now part of the start phase (function startup()). The state kStarted is
the state that in which the function run() is used and handles the
actual copy of the file. The states kCopyDone, kLocalFileClosed and
kRemoteFileClosed are part of the shutdown phase (function shutdown()).
This refactor is done so the class can be extended to support aborting
downloads in the future.

Test: Manual test: Symbol downloads still work.
Bug: part of http://b/230757105