Add download_disabled_modules_ list
This adds download_disabled_modules_ to OrbitApp and uses it. The list
contains the modules for which the user checked "Never download symbols
for this module" before. Making this list persistent is done via
QSettingsStorageManager. A entry is removed from the list when the user
manually clicks "Load Symbols" on this module (LoadSymbolsManually).

Bug http://b/231579238