Mark "Frame pointer unwinding" as experimental
We want to make the user notice, that frame pointer unwinding
needs to be treated with caution.

In this change, we change to color of the unwinding method selection
in the capture options to orange, when the user selects frame
pointer unwinding. In this case, we also add a tooltip, explaining
frame pointer unwinding being experimental and the requirement
of compiling with frame pointers.

Note that the feature is still hidden behind dev-mode.

Test: Change unwinding method in capture options.
Bug: no bug.