Initialize native unwinder in 'PeCoffInterface' and call it in 'Step'
The native PE/COFF unwinder is essentially completely implemented. This
change adds one of the final missing pieces: Initializing the native
unwinder in 'PeCoffInterface' and call its 'Step' method as a fallback
to the .debug_frame 'Step' call. Some additional data needed to be
parsed from the PE/COFF object file, which required some changes to the
test class 'PeCoffFake' to make sure that 'PeCoffInterface::Init' still
succeeds. The native unwinder is only initialized in the 64-bit case,
as we do not have an implementation for 32-bit unwinding. The latter is
still possible using .debug_frame data if present.

Tested: Unit tests pass; ran sample application with unwinding.
Bug: http://b/194768602