Handle overlapping new maps in LibunwindstackMapsImpl
This was a general problem, but it showed because Wine creates overlapping maps
as part of the process of loading a PE.

Also stop passing the load bias. We were doing that because `Maps::Add`
supported it, but it's not necessary: `libunwindstack` retrieves the load bias
as part of associating a map to an object when necessary during unwinding.

Bug: http://b/226562022

Test: Add unit tests for `LibunwindstackMaps`. Capture Trata, also with
dynamic instrumentation, and verify unwinding works.