Send ModuleUpdateEvents also for PEs mapped anonymously by Wine
This complements the logic in `orbit_object_utils::ParseMaps`, used at the
beginning of the capture.

It takes advantage of the logic in
`MapInfo::GetFileMemoryFromAnonExecMapIfPeCoffTextSection`. This makes it a bit
of a "hack on top of the hack", but I want to re-use the existing logic in
`MapInfo` to detect PEs whose text section is mapped anonymously instead of
implementing it a third time.

Bug: http://b/226562022

Test:
- Add unit test.
- Tried by capturing unaligned `triangle.exe` from the beginning with `--debug`.
  Verify that the callstacks are correct (make sense), in particular from the
  top-down view. They aren't without this change.
- Also tried by capturing Virtus from the beginning with `--debug` (or rather,
  by setting `LD_PRELOAD=libdebugger_preload.so` manually in the run script).