Also request and handle mmap_data events
In other words, also include non-executable mappings when keeping our snapshot
of the target's maps up to date during a capture.

This will be needed to detect if a new anonymous executable mapping corresponds
to a PE .text section mapped anonymously by Wine, both for unwinding purposes in
`libunwindstack` and to send the corresponding `ModuleUpdateEvent`s. As a
reminder, the logic is the following:
Given an anonymous executable mapping, consider the previous file mapping, and
consider the "group" of mappings for this file. If this is a PE, and there is no
executable file mapping for this PE (in this "group" of mappings), and the
address range where the .text section of this PE should be mapped (based on the
address of the first mapping in the "group" and on the .text section's
`VirtualAddress` and `VirtualSize`) is contained in the anonymous mapping's
address range, then we can be fairly sure that the anonymous mapping does indeed
correspond to the .text of the PE.
See also https://github.com/google/orbit/pull/3568 and
https://github.com/google/orbit/pull/3590.

Bug: http://b/226562022

Test:
- Add unit test for
  `UprobesUnwindingVisitor::Visit(uint64_t, const MmapPerfEventData&)`;
- Capture Trata, including from the beginning with `--debug`, and verify things
  are in order;
- Tried as part of the prototype for mixed unwinding, including the detection of
  new mappings that correspond to anonymously mapped PE text sections, by
  capturing unaligned `triangle.exe` from the beginning with `--debug`.