Specialized "QSpinBox" to validate multiples of "singleStep"
This prepares for a change in the "maximal raw stack to copy"
capture option. The kernel only accepts multiples of eight,
so a text edit is a challenging to explain the user.

QSpinBox already provides a step size as well as min/max values.
However, it also allows typing text directly, but the
SpinBox only validates that the input is integer and is between
min and max.

As we want to keep the "type in the text" functionality for
larger numbers, this change adds a specialized spin box,
that valides the numbers also being a multiple of the
single step value.

Test: Unit Tests
Bug: No bug