Create zoom-in/out buttons
In this PR We are creating new classes PlusButton and MinusButton just to
add symbols in the Buttons instead of using text that isn't nice enough. Also we
are renaming and using consistent kZValue for buttons and including a default button size.

Screenshot: http://screenshot/B5H8XkEbVGV7CZ4

Both buttons are reacting to picking as if the user perform a horizontal zooming from the
middle of the screen.

Bug: http://b/208451854