Use consistently timeline's width
In https://github.com/google/orbit/pull/3624 we modify the width of the
TrackContainer and the Timeline consistently to only take into
consideration the visible part, but the conversion to time was still
using the total width.

In this PR, instead we use timeline_width for time conversions, so we are
fixing 2 issues:

- Not having an invisible part of the capture behind the vertical scrollbar.

- WheelZooming functionality has again the invariant of not changing the
  elements is pointing to.

Bug: http://b/231273317

In addition we are adapting the tests to use timeline's size and
position and fixing an issue with the tests in the left of the Timeline
that was testing a middle position instead.

Test: Load a capture, use the mouse wheel, analyze if some part of the
capture is invisible.