Warn about many unwinding errors
Top-down and button-up views already report unwinding errors,
quite prominently to the user.
However, in the sampling tab, on the unwinding errors per function
are reported so far. So a user does not realize directly, that
a capture has many unwinding errors.

This change adds a notification box (similar to the code view)
to the sampling tab, if there are more than 5% unwinding errors
in the selected thread.

Screenshots:
"All threads" - with warning:
http://screenshot/ApwzyhdMGykqLNA

"All threads" - after "Acknowledge"
http://screenshot/ArpmydiqYUaKfH4

A single thread:
http://screenshot/8JduEUWPVhBxucG

Test: Take a capture and navigate the sampling tabs. Unit tests.
Bug: No bug