Add StopSource/StopToken to symbol downloads (#3663)
This adds StopSource and StopToken to the module download methods.
SftpCopyToLocalOperation checks whether a stop was requested in its copy
loop and if it was, jumps to a state that closes and deletes the partially
downloaded file. In a future commit stopping downloads will be enabled
in the following way. With the module id (file_path & build_id) get the
corresponding StopSource from OrbitApp::modules_currently_loading_
and call StopSource::RequestStop.

Test: Manual test that downloading still works
Bug http://b/230757105