Pass PC adjustment from general unwinder to PE/COFF unwinder
This is needed to make sure that disassembling for epilog detection
does have the right machine code address, otherwise it will (obviously)
fail.

This requires an interface change on the Object::Step(...) method, but
I don't see a different approach than this right now.

There is a potential optimization in the native unwinder based on the
value of the PC adjustment. If it's not 0, then we know we are not in
the innermost frame and can skip most of the logic of the native
unwinder (e.g. we know we can't be in the epilog).

Tested: Unit tests.
Bug: http://b/194768602