Use symbols_currently_loading_ differently
Beforehand symbols_currently_loading_ was the list of modules for which
symbols are currently loaded into the module (extracted from file) and
basically guarded the function LoadSymbols(). Afterwards
symbols_currently_loading_ is the list of modules for which symbol
loading overall is in progress, guarding RetrieveModuleAndLoadSymbols().
LoadSymbols() is still guarded, because it is only called from
RetrieveModuleAndLoadSymbols().

This also removes the need for an additional check (removed code).
Removing this check is in line with the optimal solution discussed in
http://b/205002963

Bug: part of http://b/231132857